% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crispr_heatmap_Module.R
\name{ClusteringServerMod}
\alias{ClusteringServerMod}
\title{Draw clusters heatmap from counts matrix server side}
\usage{
ClusteringServerMod(
  input,
  output,
  session,
  data = NULL,
  metadata = NULL,
  printRows = FALSE,
  vst = FALSE
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{data}{}

\item{name}{Does the file have a Header}
}
\description{
Draw clusters heatmap from counts matrix server side
}
