% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_app.R, R/crispr_app_profvis.R
\name{crispr_app}
\alias{crispr_app}
\title{Launch App function}
\usage{
crispr_app()

crispr_app()
}
\value{
None

None
}
\description{
Function launching the crispr app..;

Function launching the crispr app..;
}
\examples{
if (interactive()) crispr_app()
if (interactive()) crispr_app()
}
