% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_crispr_app}
\alias{server_crispr_app}
\title{Server function}
\usage{
server_crispr_app(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
None
}
\description{
Function containing all the server side of the Shiny App, must be called inside the run_app() function
}
