% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crispr_DEA_Module.R
\name{CRISPRDeaModServer}
\alias{CRISPRDeaModServer}
\title{Perform DE analysis server side}
\usage{
CRISPRDeaModServer(
  input,
  output,
  session,
  sampleplan = NULL,
  var = NULL,
  norm_data = NULL
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Perform DE analysis server side
}
